rename EmbedHelper to EmbedManager
