+++ src/main/java/Bots/commands/dev/CommandSendLogs.java
@@ -16,7 +16,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.stream.Stream;
 
-import static Bots.EmbedHelper.createQuickEmbed;
+import static Bots.managers.EmbedManager.createQuickEmbed;
 
 public class CommandSendLogs extends BaseCommand {
     @Override
