+++ src/main/java/Bots/commands/dj/CommandAutoplay.java
@@ -9,7 +9,7 @@ import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 
 import java.util.List;
 
-import static Bots.EmbedHelper.createQuickEmbed;
+import static Bots.managers.EmbedManager.createQuickEmbed;
 import static Bots.Main.*;
 import static Bots.lavaplayer.LastFMManager.filterMetadata;
