+++ src/main/java/Bots/commands/dj/CommandVibrato.java
@@ -9,7 +9,7 @@ import com.github.natanbc.lavadsp.vibrato.VibratoPcmAudioFilter;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
 
-import static Bots.EmbedHelper.createQuickEmbed;
+import static Bots.managers.EmbedManager.createQuickEmbed;
 import static Bots.Main.AudioFilters;
 
 public class CommandVibrato extends BaseCommand {
