+++ src/main/java/Bots/commands/music/CommandNowPlaying.java
@@ -13,7 +13,7 @@ import net.dv8tion.jda.api.EmbedBuilder;
 
 import java.util.Objects;
 
-import static Bots.EmbedHelper.toSimpleTimestamp;
+import static Bots.managers.EmbedManager.toSimpleTimestamp;
 import static Bots.Main.*;
 
 public class CommandNowPlaying extends BaseCommand {
