+++ src/main/java/Bots/commands/music/CommandPause.java
@@ -9,7 +9,7 @@ import Bots.lavaplayer.PlayerManager;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayer;
 
 import static Bots.CommandStateChecker.PerformChecks;
-import static Bots.EmbedHelper.createQuickEmbed;
+import static Bots.managers.EmbedManager.createQuickEmbed;
 
 public class CommandPause extends BaseCommand {
     @Override
