+++ src/main/java/Bots/lavaplayer/PlayerManager.java
@@ -37,8 +37,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import static Bots.CommandEvent.createQuickError;
-import static Bots.EmbedHelper.sanitise;
-import static Bots.EmbedHelper.toTimestamp;
+import static Bots.managers.EmbedManager.sanitise;
+import static Bots.managers.EmbedManager.toTimestamp;
 import static Bots.managers.LocaleManager.managerLocalise;
 import static Bots.Main.*;
