+++ src/main/java/Bots/lavaplayer/TrackScheduler.java
@@ -17,8 +17,8 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 
 import static Bots.CommandEvent.createQuickError;
-import static Bots.EmbedHelper.createQuickEmbed;
-import static Bots.EmbedHelper.toSimpleTimestamp;
+import static Bots.managers.EmbedManager.createQuickEmbed;
+import static Bots.managers.EmbedManager.toSimpleTimestamp;
 import static Bots.managers.LocaleManager.managerLocalise;
 import static Bots.Main.*;
 import static Bots.lavaplayer.LastFMManager.filterMetadata;
